deprecated previous cal gen script, refactored compile_info function, added format_times function, updated o365 email msg, started removing unused dependencies
